Added command line interface, stopped creating encoders and decoders anew in the training loop.
Now the same model can be trained multiple times, easing access.

A note on the command line arguments. Some of the arguments reference the source and target language. This is not used yet, but if we agree on a naming and storing convention, we can build these into the filenames of the corpora, this reducing noticeably the number of parameters (and getting rid of filenames - the most bulky ones)